pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "config-data-validator"
version = "0.1.0"
dependencies = [
    "pyyaml",
    "jsonschema",
    "click"
]

[tool.setuptools.packages.find]
where = ["src"]

